@@ -1,17 +1,27 @@
 {
   "name": "react-scrolly",
   "version": "0.1.0",
-  "workspaces": [
-    "packages/*"
-  ],
+  "workspaces": {
+    "packages": [
+      "packages/*",
+      "docs"
+    ]
+  },
   "private": true,
   "scripts": {
-    "start": "npm run clean && npm run dev",
-    "dev": "docz dev --port 5000",
-    "clean": "lerna run --parallel clean",
-    "build:doc": "docz build",
-    "prepublish": "lerna run prepublish && npm run build:doc",
-    "lint": "lerna run lint --parallel --stream",
+    "bootstrap": "lerna bootstrap",
+    "postinstall": "npm run bootstrap",
+    "clean": "lerna run clean",
+    "start": "yarn dev",
+    "dev": "run-p dev:packages dev:docs",
+    "dev:packages": "lerna run dev --parallel --stream",
+    "dev:docs": "yarn workspace docs dev",
+    "build": "lerna run build --stream",
+    "_start": "npm run clean && npm run _dev",
+    "_dev": "docz dev --port 5000",
+    "_build:doc": "docz build",
+    "lint": "eslint --ignore-path .gitignore \"./**/*.{ts,tsx}\"",
+    "lint:fix": "yarn lint --fix",
     "bumpversion": "lerna version --no-push --no-git-tag-version",
     "lerna": "lerna",
     "release": "lerna publish from-package --yes --no-push --no-git-tag-version",
@@ -20,28 +30,29 @@
     "deploy": "DOCZ_BASE=/react-scrolly/ npm run build:doc && gh-pages -d .docz/dist"
   },
   "devDependencies": {
-    "@types/node": "^10.12.18",
-    "@types/react": "^16.8.10",
-    "@types/styled-components": "4.1.8",
-    "docz": "^1.2.0",
+    "@garfieldduck/eslint-config": "^0.0.3",
+    "@garfieldduck/prettier-config": "^1.0.2",
+    "@garfieldduck/typescript-config": "^1.0.3",
+    "@types/node": "^14.0.23",
+    "@types/react": "^16.9.43",
+    "@types/styled-components": "5.1.1",
+    "docz": "^1.3.2",
     "docz-theme-default": "^1.2.0",
+    "eslint": "^6.7.2",
     "gh-pages": "^2.0.1",
     "lerna": "^3.15.0",
     "npm-run-all": "^4.1.5",
     "polish": "^0.2.3",
-    "prop-types": "^15.7.1",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "prettier": "^2.0.5",
+    "prop-types": "^15.7.2",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "react-spring": "^8.0.18",
-    "rimraf": "^2.6.3",
-    "styled-components": "^4.1.3",
-    "styled-media-query": "^2.1.1",
+    "rimraf": "^3.0.2",
+    "styled-components": "^5.1.1",
+    "styled-media-query": "^2.1.2",
     "ts-loader": "^5.3.3",
-    "tslint": "^5.12.1",
-    "tslint-config-airbnb": "^5.11.1",
-    "tslint-react": "^3.6.0",
-    "tslint-react-hooks": "^1.1.0",
-    "typescript": "^3.5.1",
+    "typescript": "^3.9.7",
     "webpack": "^4.29.6",
     "webpack-cli": "^3.3.0"
   },
@@ -50,5 +61,9 @@
   },
   "resolutions": {
     "ansi-styles": "^3.2.0"
+  },
+  "eslintConfig": {
+    "root": true,
+    "extends": "./configs/.eslintrc.js"
   }
 }
