@@ -1,16 +1,6 @@
 import { Observable, of, zip } from 'rxjs';
-import {
-  map,
-  switchMap,
-  merge,
-  take,
-} from 'rxjs/operators';
-import {
-  useState,
-  useEffect,
-  useRef,
-  useContext,
-} from 'react';
+import { map, switchMap, merge, take } from 'rxjs/operators';
+import { useState, useEffect, useRef, useContext } from 'react';
 
 import { PageContext, PageContextInterface } from '../../context/PageContext';
 import { ScrollPosition } from '../../components/PageProvider';
@@ -18,9 +8,7 @@ import { ScrollPosition } from '../../components/PageProvider';
 import { IntersectionInfo } from './useIntersectionObservable';
 import { useSubscription } from './useSubscription';
 
-export function usePageScroll(
-  intersectObsr$: Observable<IntersectionInfo>,
-) {
+export function usePageScroll(intersectObsr$: Observable<IntersectionInfo>) {
   const context = useContext<PageContextInterface | null>(PageContext);
   const { scrollObs$ } = context!;
 
@@ -35,22 +23,19 @@ export function usePageScroll(
 
   /** Function to set the subscription to the page scrolling  */
   const { setSubscription: setPageSubscpt } = useSubscription(null);
-  const isIntersectingObs = useRef(intersectObsr$.pipe(
-    map(({ isIntersecting }) => isIntersecting)
-  ));
+  const isIntersectingObs = useRef(
+    intersectObsr$.pipe(map(({ isIntersecting }) => isIntersecting))
+  );
 
   /**
    * Observer to track the scroll position
    * emitted when Page is mounted
    */
   const mountScrollObsRef = useRef(
-    zip(
-      scrollObs$,
-      isIntersectingObs.current,
-    ).pipe(
+    zip(scrollObs$, isIntersectingObs.current).pipe(
       map(([scrollPos, isIntersecting]): {
-        isIntersecting: boolean,
-        scrollPos: ScrollPosition,
+        isIntersecting: boolean;
+        scrollPos: ScrollPosition;
       } => ({
         isIntersecting,
         scrollPos: {
@@ -58,7 +43,7 @@ export function usePageScroll(
           scrollOffset: 0,
         },
       })),
-      take(1),
+      take(1)
     )
   );
   /**
@@ -71,30 +56,28 @@ export function usePageScroll(
         // use `isIntersecting` to determine whether to take the scrolling info
         return isIntersecting
           ? scrollObs$.pipe(
-            map((scrollPos: ScrollPosition) => ({
+              map((scrollPos: ScrollPosition) => ({
+                isIntersecting,
+                scrollPos,
+              }))
+            )
+          : // when the section is scrolled out of the viewport, update its dimension
+            of({
               isIntersecting,
-              scrollPos,
-            })),
-          )
-          // when the section is scrolled out of the viewport, update its dimension
-          : of({
-            isIntersecting,
-            scrollPos: null,
-          });
+              scrollPos: null,
+            });
       })
     )
   );
   /** Observer to track the page scrolling by combining mountScrollObs and windowScrollObs */
   const pageScrollObsrRef = useRef(
-    mountScrollObsRef.current.pipe(
-      merge(windowScrollObsRef.current),
-    )
+    mountScrollObsRef.current.pipe(merge(windowScrollObsRef.current))
   );
 
-  useEffect(
-    () => {
-      // subscribe to the page scrolling
-      setPageSubscpt(pageScrollObsrRef.current.subscribe({
+  useEffect(() => {
+    // subscribe to the page scrolling
+    setPageSubscpt(
+      pageScrollObsrRef.current.subscribe({
         // record the page scrolling
         next: ({ isIntersecting, scrollPos }) => {
           if (isIntersecting && scrollPos) {
@@ -112,10 +95,9 @@ export function usePageScroll(
           // update the intersecting state
           setIntersecting(isIntersecting);
         },
-      }));
-    },
-    [],
-  );
+      })
+    );
+  }, [setPageSubscpt]);
 
   return {
     scrollInfo,
