@@ -1,17 +1,6 @@
 import { Observable, of } from 'rxjs';
-import {
-  map,
-  filter,
-  switchMap,
-  merge,
- } from 'rxjs/operators';
-import React, {
-  useState,
-  useEffect,
-  useRef,
-  useContext,
-  useCallback,
-} from 'react';
+import { map, filter, switchMap, merge } from 'rxjs/operators';
+import React, { useState, useEffect, useRef, useContext, useCallback } from 'react';
 
 import { PageContext, PageContextInterface } from '../../context/PageContext';
 
@@ -29,9 +18,8 @@ export interface SectionPosition {
 export function useSectionPosition(
   /** Ref of the section being tracked */
   sectionRef: React.RefObject<HTMLElement>,
-  intersectObsr$: Observable<IntersectionInfo>,
+  intersectObsr$: Observable<IntersectionInfo>
 ): SectionPosition {
-
   const context = useContext<PageContextInterface | null>(PageContext);
   const { resizeObs$ } = context!;
 
@@ -51,38 +39,42 @@ export function useSectionPosition(
   const { setSubscription } = useSubscription(null);
 
   /** Observer to the window resizing events */
-  const combinedResizeObsRef = useRef(intersectObsr$.pipe(
-    switchMap((intersectInfo) => {
-      const { isIntersecting, sectionBoundingRect } = intersectInfo;
-      return isIntersecting
-        ? (
-          // return the bounding rect from `intersectObsr$` when it appears in the viewport
-          of(sectionBoundingRect)
-          // merge it with the intersection observer
-          .pipe(
-            merge(resizeObs$.pipe(
-              map(() => {
-                const currentSect = sectionRef.current;
-                if (currentSect) {
-                  const rect = currentSect.getBoundingClientRect();
-                  return rect;
-                }
-                return undefined;
-              }),
-            ))
-          )
+  const combinedResizeObsRef = useRef(
+    resizeObs$
+      ? intersectObsr$.pipe(
+          switchMap((intersectInfo) => {
+            const { isIntersecting, sectionBoundingRect } = intersectInfo;
+            return isIntersecting
+              ? // return the bounding rect from `intersectObsr$` when it appears in the viewport
+                of(sectionBoundingRect)
+                  // merge it with the intersection observer
+                  .pipe(
+                    merge(
+                      resizeObs$.pipe(
+                        map(() => {
+                          const currentSect = sectionRef.current;
+                          if (currentSect) {
+                            const rect = currentSect.getBoundingClientRect();
+                            return rect;
+                          }
+                          return undefined;
+                        })
+                      )
+                    )
+                  )
+              : // when the section is scrolled out of the viewport
+                of(undefined);
+          }),
+          filter((rect) => typeof rect !== 'undefined')
         )
-        // when the section is scrolled out of the viewport
-        : of(undefined);
-    }),
-    filter(rect => typeof rect !== 'undefined'),
-  ));
+      : undefined
+  );
 
   /**
    * Update the absolute position of the section
    */
-  const updateSectionPosition = useCallback(
-    (boundingClientRect) => {
+  const updateSectionPosition = useCallback((boundingClientRect) => {
+    if (typeof window !== 'undefined') {
       const scrollTop = window.pageYOffset || document.documentElement.scrollTop;
       const { top } = boundingClientRect;
 
@@ -90,26 +82,27 @@ export function useSectionPosition(
         sectionTop: top + scrollTop,
         boundingRect: boundingClientRect,
       });
-    },
-    [],
-  );
+    }
+  }, []);
 
-  useEffect(
-    () => {
-      // update the dimension of the section when it's mounted
-      if (sectionRef.current) {
-        updateSectionPosition(sectionRef.current.getBoundingClientRect());
-      }
+  useEffect(() => {
+    // update the dimension of the section when it's mounted
+    if (sectionRef.current) {
+      updateSectionPosition(sectionRef.current.getBoundingClientRect());
+    }
 
-      // set the subscription to the page resizing events and intersection events
-      setSubscription(combinedResizeObsRef.current.subscribe({
-        next: (sectionBoundingRect) => {
-          updateSectionPosition(sectionBoundingRect!);
-        },
-      }));
-    },
-    [],
-  );
+    // set the subscription to the page resizing events and intersection events
+    // combinedResizeObsRef.current does not exist on SSR
+    if (combinedResizeObsRef.current) {
+      setSubscription(
+        combinedResizeObsRef.current.subscribe({
+          next: (sectionBoundingRect) => {
+            updateSectionPosition(sectionBoundingRect!);
+          },
+        })
+      );
+    }
+  }, [sectionRef, setSubscription, updateSectionPosition]);
 
   return sectionPosition;
 }
