@@ -3,7 +3,6 @@ import styledComponents from 'styled-components';
 export const DemoWrapper = styledComponents.div`
   padding: 2rem 1rem;
   margin: 1.8rem 0;
-  background: #ffffff;
   border-radius: 5px;
-  box-shadow: 0 12px 22px 0 rgba(0,0,0,.06), 0 6px 14px 0 rgba(0,0,0,.03);
+  box-shadow: 0 12px 22px 0 rgba(180,180,180,.3), 0 6px 14px 0 rgba(180,180,180,.3);
 `;
